+++ util/clixon_util_path.c
@@ -241,7 +241,7 @@ main(int    argc,
 	if (xml_apply0(x, -1, xml_sort_verify, h) < 0)
 	    clicon_log(LOG_NOTICE, "%s: sort verify failed", __FUNCTION__);
 	/* Add default values */
-	if (xml_default_recurse(x) < 0)
+	if (xml_default_recurse(x, 0) < 0)
 	    goto done;
 	if ((ret = xml_yang_validate_all_top(h, x, &xerr)) < 0) 
 	    goto done;