+++ CHANGELOG.md
@@ -46,7 +46,13 @@ Users may have to change how they access the system
 
 ### Minor changes
 
+* 
 * Added sanity check that a yang module name matches the filename
+
+### Corrected Bugs
+
+* Fixed: [default state data returned with get-config](https://github.com/clicon/clixon/issues/140)
+  * Generalized default code for both config and state
 
 ## 4.7.0
 14 September 2020